Refactor: Use dependencyManagement, declare slf4j, logback only in runtime scope
