+++ tcases-lib/pom.xml
@@ -1,4 +1,4 @@
-<project 
+<project
     xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
@@ -56,7 +56,7 @@
       <subscribe>https://groups.google.com/d/forum/tcases</subscribe>
       <unsubscribe>https://groups.google.com/d/forum/tcases</unsubscribe>
       <post>https://groups.google.com/d/forum/tcases</post>
-    </mailingList>  
+    </mailingList>
   </mailingLists>
 
   <properties>
@@ -79,7 +79,6 @@
         <filtering>true</filtering>
       </testResource>
     </testResources>
-
   </build>
 
   <dependencies>
@@ -87,40 +86,35 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
     </dependency>
 
     <dependency>
       <groupId>commons-io</groupId>
       <artifactId>commons-io</artifactId>
-      <version>2.4</version>
+      <scope>test</scope>
     </dependency>
 
     <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
-      <version>3.3.1</version>
     </dependency>
 
     <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-collections4</artifactId>
-      <version>4.1</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
     </dependency>
 
     <dependency>
       <groupId>ch.qos.logback</groupId>
       <artifactId>logback-classic</artifactId>
-      <version>1.1.1</version>
+      <scope>test</scope>
     </dependency>
 
-    <dependency>
-      <groupId>org.apache.ant</groupId>
-      <artifactId>ant</artifactId>
-      <version>1.8.2</version>
-    </dependency>
-    
   </dependencies>
 
 </project>
