+++ tcases-lib/src/main/java/org/cornutum/tcases/Reducer.java
@@ -11,7 +11,7 @@ import org.cornutum.tcases.generator.*;
 import org.cornutum.tcases.generator.io.*;
 import org.cornutum.tcases.io.*;
 import org.apache.commons.collections4.IteratorUtils;
-import org.apache.commons.io.IOUtils;
+import org.cornutum.tcases.util.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
