+++ tcases-lib/src/main/java/org/cornutum/tcases/io/AbstractFilter.java
@@ -7,7 +7,7 @@
 
 package org.cornutum.tcases.io;
 
-import org.apache.commons.io.IOUtils;
+import org.cornutum.tcases.util.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
