+++ tcases-lib/src/test/java/org/cornutum/tcases/TestTcasesCommand.java
@@ -9,7 +9,7 @@ package org.cornutum.tcases;
 
 import org.cornutum.tcases.Tcases.Options;
 import org.cornutum.tcases.io.SystemTestResources;
-
+import static org.cornutum.tcases.TcasesCommand.OptionsFactory.parseOptions;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
@@ -26,13 +26,13 @@ import java.net.URL;
 import java.nio.charset.Charset;
 
 /**
- * Runs tests for {@link Tcases#main}.
+ * Runs tests for {@link TcasesCommand#main}.
  *
  */
-public class TestTcases
+public class TestTcasesCommand
   {
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 0. run (Success) </TH></TR>
@@ -81,14 +81,14 @@ public class TestTcases
       };
     
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFile.exists());
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 1. run (Success) </TH></TR>
@@ -134,14 +134,14 @@ public class TestTcases
       };
 
     // When...
-    runWithStdIO( new Options( args), inFile, null);
+    runWithStdIO( parseOptions( args), inFile, null);
         
     // Then...
     assertEquals( "Test def created", true, testDefFile.exists());
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 2. run (Success) </TH></TR>
@@ -185,7 +185,7 @@ public class TestTcases
       };
 
     // When...
-    runWithStdIO( new Options( args), inFile, null);
+    runWithStdIO( parseOptions( args), inFile, null);
         
     // Then...
     File outFile = new File( outDir, testDefFile.getName());
@@ -193,7 +193,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 3. run (Success) </TH></TR>
@@ -236,18 +236,18 @@ public class TestTcases
         "run-3"
       };
 
-    Options options = new Options( args);
+    Options options = parseOptions( args);
     options.setWorkingDir( inFile.getParentFile());
     
     // When...
-    Tcases.run( options);
+    TcasesCommand.run( options);
         
     // Then...
     assertEquals( "Test def created", true, outFile.exists());
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 4. run (Success) </TH></TR>
@@ -292,14 +292,14 @@ public class TestTcases
       };
 
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFile.exists());
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 5. run (Success) </TH></TR>
@@ -341,14 +341,14 @@ public class TestTcases
 
     // When...
     StringBuffer outFile = new StringBuffer();
-    runWithStdIO( new Options( args), inFile, outFile);
+    runWithStdIO( parseOptions( args), inFile, outFile);
         
     // Then...
     assertEquals( "Test def created", true, outFile.length() > 0);
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 6. run (Failure) </TH></TR>
@@ -396,7 +396,7 @@ public class TestTcases
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -406,7 +406,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 7. run (Failure) </TH></TR>
@@ -443,13 +443,13 @@ public class TestTcases
         "run-7"
       };
 
-    Options options = new Options( args);
+    Options options = parseOptions( args);
     options.setWorkingDir( inFile.getParentFile());
 
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -459,7 +459,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 8. run (Failure) </TH></TR>
@@ -500,7 +500,7 @@ public class TestTcases
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -510,7 +510,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 9. run (Failure) </TH></TR>
@@ -555,9 +555,9 @@ public class TestTcases
     // When...
     try
       {
-      Options options = new Options( args);
+      Options options = parseOptions( args);
       options.setWorkingDir( inFile.getParentFile());
-      Tcases.run( options);
+      TcasesCommand.run( options);
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -567,7 +567,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 10. run (Success) </TH></TR>
@@ -613,7 +613,7 @@ public class TestTcases
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -623,7 +623,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 0. run (Success) </TH></TR>
@@ -665,14 +665,14 @@ public class TestTcases
       };
     
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFilePath.exists());
     }
   
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 1. run (Success) </TH></TR>
@@ -709,14 +709,14 @@ public class TestTcases
 
     // When...
     StringBuffer outFile = new StringBuffer();
-    runWithStdIO( new Options( args), inFile, outFile);
+    runWithStdIO( parseOptions( args), inFile, outFile);
         
     // Then...
     assertEquals( "Test def created", true, outFile.length() > 0);
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 2. run (Success) </TH></TR>
@@ -760,14 +760,14 @@ public class TestTcases
       };
     
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFilePath.exists());
     }
   
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 3. run (Success) </TH></TR>
@@ -806,14 +806,14 @@ public class TestTcases
       };
 
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFilePath.exists());
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 4. run (<FONT color="red">Failure</FONT>) </TH></TR>
@@ -853,7 +853,7 @@ public class TestTcases
     Exception failure = null;
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       }
     catch( Exception expected)
       {
@@ -869,7 +869,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 5. run (<FONT color="red">Failure</FONT>) </TH></TR>
@@ -907,7 +907,7 @@ public class TestTcases
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -917,7 +917,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 6. run (<FONT color="red">Failure</FONT>) </TH></TR>
@@ -957,7 +957,7 @@ public class TestTcases
     Exception failure = null;
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       }
     catch( Exception expected)
       {
@@ -973,7 +973,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 7. run (<FONT color="red">Failure</FONT>) </TH></TR>
@@ -1016,7 +1016,7 @@ public class TestTcases
     // When...
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       fail( "Expected exception not thrown");
       }
     // Then...
@@ -1026,7 +1026,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 8. run (<FONT color="red">Failure</FONT>) </TH></TR>
@@ -1066,7 +1066,7 @@ public class TestTcases
     Exception failure = null;
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       }
     catch( Exception expected)
       {
@@ -1092,7 +1092,7 @@ public class TestTcases
     failure = null;
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       }
     catch( Exception expected)
       {
@@ -1118,7 +1118,7 @@ public class TestTcases
     failure = null;
     try
       {
-      Tcases.run( new Options( args));
+      TcasesCommand.run( parseOptions( args));
       }
     catch( Exception expected)
       {
@@ -1134,7 +1134,7 @@ public class TestTcases
     }
 
   /**
-   * Tests {@link Tcases#run run()} using the following inputs.
+   * Tests {@link TcasesCommand#run run()} using the following inputs.
    * <P>
    * <TABLE border="1" cellpadding="8">
    * <TR align="left"><TH colspan=2> 9. run (Success) </TH></TR>
@@ -1172,7 +1172,7 @@ public class TestTcases
       };
     
     // When...
-    Tcases.run( new Options( args));
+    TcasesCommand.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFilePath.exists());
@@ -1193,7 +1193,7 @@ public class TestTcases
       };
     
     // When...
-    Tcases.run( new Options( args));
+    Tcases.run( parseOptions( args));
         
     // Then...
     assertEquals( "Test def created", true, outFilePath.exists());
@@ -1206,7 +1206,7 @@ public class TestTcases
     InputStream inputDef = getClass().getResourceAsStream( "tcases-Transform-Input.xml");
     
     // When...
-    SystemTestDef testDef = Tcases.getTests( inputDef);
+    SystemTestDef testDef = TcasesCommand.getTests( inputDef);
     
     // Then...
     SystemTestDef expectedTestDef = testResources_.read( "tcases-Transform-Test.xml");
@@ -1214,7 +1214,7 @@ public class TestTcases
 
     // When...
     ByteArrayOutputStream testDefOut = new ByteArrayOutputStream();
-    Tcases.writeTests( testDef, testDefOut);
+    TcasesCommand.writeTests( testDef, testDefOut);
 
     // Then...
     ByteArrayInputStream testDefString = new ByteArrayInputStream( testDefOut.toByteArray());
@@ -1229,7 +1229,7 @@ public class TestTcases
     InputStream genDef = getClass().getResourceAsStream( "tcases-Transform-Generators.xml");
     
     // When...
-    SystemTestDef testDef = Tcases.getTests( inputDef, genDef, null);
+    SystemTestDef testDef = TcasesCommand.getTests( inputDef, genDef, null);
     
     // Then...
     SystemTestDef expectedTestDef = testResources_.read( "tcases-Transform-Gen-Test.xml");
@@ -1245,7 +1245,7 @@ public class TestTcases
     InputStream baseDef = getClass().getResourceAsStream( "tcases-Transform-Test.xml");
     
     // When...
-    SystemTestDef testDef = Tcases.getTests( inputDef, genDef, baseDef);
+    SystemTestDef testDef = TcasesCommand.getTests( inputDef, genDef, baseDef);
     
     // Then...
     SystemTestDef expectedTestDef = testResources_.read( "tcases-Transform-Regen-Test.xml");
@@ -1288,7 +1288,7 @@ public class TestTcases
         System.setOut( (newOut = new PrintStream( (newOutBytes = new ByteArrayOutputStream()))));
         }
 
-      Tcases.run( options);
+      TcasesCommand.run( options);
       }
     catch( Exception e)
       {
