+++ tcases-shell/docs/Tcases-Guide.htm
@@ -279,7 +279,7 @@
       <P>
         For details about the interface to the <SPAN class="code">tcases</SPAN> command (and
         the <SPAN class="code">tcases.bat</SPAN> command, too), see the Javadoc for
-        the <A href="api/org/cornutum/tcases/Tcases.Options.html"><SPAN class="code">Tcases.Options</SPAN></A>
+        the <A href="api/org/cornutum/tcases/TcasesCommand.OptionsFactory.html"><SPAN class="code">Tcases.Options</SPAN></A>
         class.  To get help at the command line, run <SPAN class="code">tcases -help</SPAN>.
       </P>
 
@@ -1604,7 +1604,7 @@
       </P>
       <P>
         Of course, you can use various options for the <SPAN class="code">tcases</SPAN> command to customize this default
-        pattern. For details, see the <A href="api/org/cornutum/tcases/Tcases.Options.html"><SPAN class="code">Tcases.Options</SPAN></A> class,
+        pattern. For details, see the <A href="api/org/cornutum/tcases/TcasesCommand.OptionsFactory.html"><SPAN class="code">Tcases.Options</SPAN></A> class,
         or run <SPAN class="code">tcases -help</SPAN>.
       </P>
 
@@ -2456,7 +2456,7 @@ public class Examples { <BR/>
 
         <UL>
           <LI> The <SPAN class="code">find</SPAN> command <A href="examples/find-Input.xml">example</A></LI><P/>
-          <LI> The <SPAN class="code">tcases</SPAN> <A href="api/org/cornutum/tcases/Tcases.Options.html">command line</A></LI><P/>
+          <LI> The <SPAN class="code">tcases</SPAN> <A href="api/org/cornutum/tcases/TcasesCommand.OptionsFactory.html">command line</A></LI><P/>
           <LI> Document schemas <P/>
             <UL>
               <LI> The <A href="system-input.xsd">system input definition document schema</A> </LI><P/>
